Update Quadratic_Formula.py
Now takes decimal values. Changed int to float.

+++ Quadratic_Formula.py
@@ -1,16 +1,16 @@
 def solve(x, y, z, lis):
     w = int_check(lis)
     if w:
-        minus_b = int(y) * int(-1)
+        minus_b = float(y) * float(-1)
 #        print('minus_b = ' + str(minus_b))
-        b_squared = pow(int(y), 2.0)
+        b_squared = pow(float(y), 2.0)
 #        print('b_squared = ' + str(b_squared))
-        four_a_c = 4 * int(x) * int(z)
+        four_a_c = 4 * float(x) * float(z)
 #        print('four_a_c = ' + str(four_a_c))
-        two_a = 2 * int(x)
+        two_a = 2 * float(x)
 #        print('two_a = ' + str(two_a))
         to_be_square_root = b_squared - four_a_c
-        square_root = pow(int(to_be_square_root), 0.5)
+        square_root = pow(float(to_be_square_root), 0.5)
         var1 = minus_b + square_root
         var2 = minus_b - square_root
         var3 = var1/two_a
@@ -36,7 +36,7 @@ abc = [a, b, c]
 def int_check(lis):
     for i in lis:
         try:
-            int(i)
+            float(i)
             return True
         except ValueError:
             return False
